Add mock tests
