+++ src/lib.rs
@@ -28,4 +28,3 @@ mod value;
 pub use connection::*;
 pub use error::*;
 pub use value::*;
-pub use bindings::*;
